requirements.txt
streamlit>=1.28.0
requests>=2.31.0
psycopg2-binary>=2.9.9
beautifulsoup4>=4.12.0
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0
python-dotenv>=1.0.0
